refactored refresh button into its own component, deleted unnecessary route on redux
