+++ webapp/src/components/Breadcrumb/style/breadcrumb.scss
@@ -2,8 +2,6 @@
 @import "../../../styles//fonts.scss";
 
 .directory-container {
-    padding: 0 20px;
-    
     & .title {
         margin-left: 2vw;
         margin-bottom: 0;
@@ -15,8 +13,6 @@
     display: flex;
     align-items: center;
     justify-content: left;
-    border-top: 2px solid $soft-gray;
-    padding: 20px 5vw;
 }
 
 .route-breadcrumb {
@@ -32,31 +28,12 @@
     width: 100%;
 }
 
-.refresh-btn {
-    border-radius: 50px;
-    padding: 5px 20px;
-    border: 1px solid $orange;
-    font-family: $font-asap;
-}
-
-.refresh-btn:hover {
-    background-color: $volume-orange;
-    color: white;
-}
-
-.refresh-btn:active {
-    background-color: $orange-active;
-    color: white;
-}
-
 .route-breadcrumb:hover {
     color: $orange;
 }
 
 @media (min-width: 1024px) {
     .directory-container {
-        padding: 0 20px;
-
         & .title { 
             margin-left: 1vw;
         }
