+++ webapp/src/components/RefreshButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './RefreshButton.container';
