+++ webapp/src/store/reducers/index.js
@@ -3,7 +3,6 @@ import { combineReducers } from 'redux';
 import appReducer from './app';
 import filesReducer from './files';
 import slidebar from './slideBar';
-import routeReducer from './route';
 import songIndex from './songIndex';
 import songsQueue from './songsQueue';
 import auth from './auth';
@@ -17,7 +16,6 @@ export default combineReducers({
   app: appReducer,
   files: filesReducer,
   slidebarIndex: slidebar,
-  route: routeReducer,
   songIndex,
   songsQueue,
   auth,
