--- webapp/src/store/reducers/route.js
@@ -1,11 +1,0 @@
-const routeReducer = (state = '/', { type, payload = '' }) => {
-  switch (type) {
-    case 'CHANGE_ROUTE':
-      return payload;
-
-    default:
-      return state;
-  }
-};
-
-export default routeReducer;
